+++ R/emapplot_utilities.R
@@ -193,6 +193,7 @@ merge_compareClusterResult <- function(yy) {
 build_ggraph <- function(x, enrichDf, mergedEnrichDf, cex_category, pie, 
                          layout, coords, cex_line, min_edge, pair_sim,
                          method, with_edge){
+    segment.size <- get_ggrepel_segsize()
     geneSets <- setNames(strsplit(as.character(mergedEnrichDf$geneID), "/",
                               fixed = TRUE), mergedEnrichDf$ID) 
                               
@@ -221,7 +222,7 @@ build_ggraph <- function(x, enrichDf, mergedEnrichDf, cex_category, pie,
                 cols=names(ID_Cluster_mat)[1:(ncol(ID_Cluster_mat)-3)],
                 color=NA)+
             xlim(-3,3) + ylim(-3,3) + coord_equal()+
-            geom_node_text(aes_(label=~name), repel=TRUE) +
+            geom_node_text(aes_(label=~name), repel=TRUE, segment.size = segment.size) +
             theme_void()+labs(fill = "Cluster")
         return(p)
 
@@ -363,6 +364,7 @@ get_label_location <- function(ggData, label_format) {
 add_group_label <- function(label_style, repel, shadowtext, p, label_location, 
                             label_group, cex_label_group, ...) {
     if (label_style != "shadowtext") return(p)
+    segment.size <- get_ggrepel_segsize()
     if (!repel) {
         if (shadowtext) {
             p <- p + geom_shadowtext(data = label_location,
@@ -381,12 +383,12 @@ add_group_label <- function(label_style, repel, shadowtext, p, label_location,
         p <- p + ggrepel::geom_text_repel(data = label_location,
             aes_(x =~ x, y =~ y, label =~ label), colour = "black",
             size = label_group * cex_label_group, bg.color = "white", bg.r = 0.1,
-            show.legend = FALSE, ...)
+            show.legend = FALSE, segment.size = segment.size, ...)
     } else {
         p <- p + ggrepel::geom_text_repel(data = label_location,
             aes_(x =~ x, y =~ y, label =~ label), colour = "black",
             size = label_group * cex_label_group, 
-            show.legend = FALSE, ...)
+            show.legend = FALSE, segment.size = segment.size, ...)
     }
     return(p)   
 }
@@ -401,12 +403,15 @@ add_group_label <- function(label_style, repel, shadowtext, p, label_location,
 ##' @return a ggplot2 object.
 ##' @noRd
 add_node_label <- function(p, data, label_size_node, cex_label_node, shadowtext) {
+    segment.size <- get_ggrepel_segsize()
     if (shadowtext) {
         p <- p + geom_node_text(aes_(label=~name), data = data,
-            size = label_size_node * cex_label_node, bg.color = "white", repel=TRUE)
+            size = label_size_node * cex_label_node, bg.color = "white", 
+            repel=TRUE, segment.size = segment.size)
     } else {
         p <- p + geom_node_text(aes_(label=~name), data = data,
-            size = label_size_node * cex_label_node, repel=TRUE)
+            size = label_size_node * cex_label_node, repel=TRUE,
+            segment.size = segment.size)
     }
     return(p)
 }